requirements.txt
streamlit
pandas
numpy
supabase
matplotlib
plotly
python-dotenv